image-to-cnc/requirements.txt
torch
torchvision
numpy
tqdm
Pillow
svgwrite
svgpathtools
scikit-learn
cairosvg 